update/ 달력에 유저정보 받아오기 및 유저 모든 데이터 가져오기 #3
