+++ src/app/page.tsx
@@ -1,11 +1,15 @@
+import { getUserId } from "@/services/home-services";
+import HomeCalendar from "./(home)/components/home-calendar";
 import HomeDate from "./(home)/components/home-date";
-import HomeDiary from "./(home)/components/home-diary";
+//import HomeDiary from "./(home)/components/home-diary";
 
 const HomePage = async () => {
+  const userId: string | undefined = await getUserId();
   return (
     <div>
       <HomeDate />
-      <HomeDiary />
+      <HomeCalendar userId={userId} />
+      {/* <HomeDiary /> */}
     </div>
   );
 };
